feat(kuma-cp) move kuma-api errors to core/rest/erors
