+++ app/kumactl/pkg/errors/formatter_test.go
@@ -2,7 +2,7 @@ package errors_test
 
 import (
 	kumactl_errors "github.com/Kong/kuma/app/kumactl/pkg/errors"
-	"github.com/Kong/kuma/pkg/api-server/types"
+	"github.com/Kong/kuma/pkg/core/rest/errors/types"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
