+++ pkg/api-server/dataplane_overview_ws.go
@@ -11,6 +11,7 @@ import (
 	"github.com/Kong/kuma/pkg/core/resources/manager"
 	"github.com/Kong/kuma/pkg/core/resources/model/rest"
 	"github.com/Kong/kuma/pkg/core/resources/store"
+	rest_errors "github.com/Kong/kuma/pkg/core/rest/errors"
 	"github.com/emicklei/go-restful"
 )
 
@@ -39,13 +40,13 @@ func (r *overviewWs) inspectDataplane(request *restful.Request, response *restfu
 
 	overview, err := r.fetchOverview(request.Request.Context(), name, meshName)
 	if err != nil {
-		HandleError(response, err, "Could not retrieve a dataplane overview")
+		rest_errors.HandleError(response, err, "Could not retrieve a dataplane overview")
 		return
 	}
 
 	res := rest.From.Resource(overview)
 	if err := response.WriteAsJson(res); err != nil {
-		HandleError(response, err, "Could not retrieve a dataplane overview")
+		rest_errors.HandleError(response, err, "Could not retrieve a dataplane overview")
 	}
 }
 
@@ -74,7 +75,7 @@ func (r *overviewWs) inspectDataplanes(request *restful.Request, response *restf
 	meshName := request.PathParameter("mesh")
 	overviews, err := r.fetchOverviews(request.Request.Context(), meshName)
 	if err != nil {
-		HandleError(response, err, "Could not retrieve dataplane overviews")
+		rest_errors.HandleError(response, err, "Could not retrieve dataplane overviews")
 		return
 	}
 
@@ -83,7 +84,7 @@ func (r *overviewWs) inspectDataplanes(request *restful.Request, response *restf
 
 	restList := rest.From.ResourceList(&overviews)
 	if err := response.WriteAsJson(restList); err != nil {
-		HandleError(response, err, "Could not list dataplane overviews")
+		rest_errors.HandleError(response, err, "Could not list dataplane overviews")
 	}
 }
