+++ pkg/core/rest/errors/types/error.go
